style: разделение на модули
